+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 - Fix multi monitor issues
 - Show monitor name in output of `list-groups` subcommand
 - Respect `workspace_auto_back_and_forth` setting (fixes #9)
+- Deprecate `workspace-back-and-forth` and `move-to-back-and-forth` now that
+  built-in i3 commands should work